% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOplot.R
\name{GO_plot}
\alias{GO_plot}
\title{A simple function for Gene Ontology enrichment analysis and plotting.}
\usage{
GO_plot(gene, orgdb = "org.Hs.eg.db")
}
\arguments{
\item{gene}{genelist from DEG}

\item{orgdb}{Genome annotation database of the species,the default is "org.Hs.eg.db".}
}
\value{
ggplot2 plot object
}
\description{
A simple function for Gene Ontology enrichment analysis and plotting.
}
\examples{
library(GOqt)
library(tidyverse)
library(clusterProfiler)
genelist = c(93650,135,57538,27063)
GO_plot(genelist)+ggtitle("GO enrichment of diff_genes")
}
\author{
qinti
}
